requirements.txt
python-dotenv==1.0.1
pyotp==2.9.0
selenium==4.18.1
